crams-apps/crams_provision/requirements.txt
pytz
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-model-utils==4.1.
merc_common==1.0.0
crams_contact==1.0.0
crams-log==1.0.0
crams_collection==1.0.0
crams-compute==1.0.0
crams-storage==1.0.0
crams_allocation==1.0.0
mysqlclient==2.0.1
rest_condition==1.0.3
gunicorn==20.0.4
urllib3==1.25.10
sqlparse==0.4.1